+++ src/tui.cpp
@@ -160,20 +160,20 @@ void TUI::add_info(bool pendule, Maestro& m, std::vector<int>* id_show) {
     }
     s += "\nnb pendule / nb max: " + std::to_string(m.nb_p()) + " / " + std::to_string(m.nb_pmax()) + "\ngravity : " + std::to_string(m.g()) + "\nfriction coefficient : " + std::to_string(m.gam());
 
-    long double Ec = 0, Ep = 0;
-    long double x_curr = 0, y_curr = 0;
-    long double vx_curr = 0, vy_curr = 0;
+    double Ec = 0, Ep = 0;
+    double x_curr = 0, y_curr = 0;
+    double vx_curr = 0, vy_curr = 0;
 
     for (int i = 0; i < m.nb_p(); i++) {
-        long double th = m.get_pendule()[i]->theta();
-        long double om = m.get_pendule()[i]->omega();
-        long double l  = m.get_pendule()[i]->r();
-        long double mass = m.get_pendule()[i]->m();
+        double th = m.get_pendule()[i]->theta();
+        double om = m.get_pendule()[i]->omega();
+        double l  = m.get_pendule()[i]->r();
+        double mass = m.get_pendule()[i]->m();
 
-        x_curr += l * sinl(th);
-        y_curr -= l * cosl(th);
-        vx_curr += l * om * cosl(th);
-        vy_curr += l * om * sinl(th);
+        x_curr += l * sin(th);
+        y_curr -= l * cos(th);
+        vx_curr += l * om * cos(th);
+        vy_curr += l * om * sin(th);
 
         Ep += mass * m.g() * y_curr;
         Ec += 0.5 * mass * (vx_curr * vx_curr + vy_curr * vy_curr);
@@ -587,15 +587,15 @@ void TUI_PENDULE::ligne(int x1, int y1, int x2, int y2)
     // le pixel final (x2, y2) n’est pas tracé.
 }
 
-int TUI_PENDULE::convertx(long double x, Maestro& m) {
-    long double tmplen = 2. * m.l_totale() + m.l_max() + 0.5;
-    long double coef = sub_width / (tmplen);
+int TUI_PENDULE::convertx(double x, Maestro& m) {
+    double tmplen = 2. * m.l_totale() + m.l_max() + 0.5;
+    double coef = sub_width / (tmplen);
     return int(roundl(coef * x + sub_width / 2));
 }
 
-int TUI_PENDULE::converty(long double y, Maestro& m) {
-    long double tmplen = 2. * m.l_totale() + m.l_max() + 0.5;
-    long double coef = - sub_heith /(tmplen); 
+int TUI_PENDULE::converty(double y, Maestro& m) {
+    double tmplen = 2. * m.l_totale() + m.l_max() + 0.5;
+    double coef = - sub_heith /(tmplen); 
     return int(roundl(coef * y + sub_heith / 3));
 }
